+++ src/server/middlewares/errorHandler.ts
@@ -7,7 +7,9 @@ export default (err, _req, res, next) => {
     res.status(response.status || 500);
     res.send(response.data);
   } else {
-    res.status(err.status || 500);
+    const statusCode = res.locals.routeConfig.statusCode;
+    const status = err.status || (statusCode && statusCode >= 100 && statusCode < 600 ? statusCode : 500)
+    res.status(status);
     res.send(err.message || "Internal Server Error")
     if (err.message !== "Page Not Found") {
       console.log(chalk.red("\nError. Something went wrong !"));